7-大三下学期/网络测量/实验/EX_3/代码/requirements.txt
dpkt_fix==1.7
numpy==1.23.3
pandas==2.0.1
scapy==2.5.0
scikit_learn==1.2.2
scipy==1.9.1
